backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-multipart
ebooklib
beautifulsoup4
pymupdf
python-dotenv